rest/reqhelper: add timing metrics to rest calls
This measures the duration of REST calls on the LINSTOR Controller side.
In order to prevent too many metric labels, we do not measure what peer
issues the request.

There are some internal calls that are not tracked. This was deemed worth it,
as tracking those would require some more invasive changes to request handling.