do not calculate timezone offset by hand
The CronTests started to fail once we switched back to CET instead of CEST.
This was caused by incorrect timezone calculations: the LOCAL_ZONE_OFFSET
was calculated based on the current time.

However, the actual offset depends both on the timezone, as well as the
date. For example, on this developers computer, the tests ran as CET, but
the test data was CEST. having an offset of +2 instead of +1.

To fix this, do not calculate the offsets at all: just use the default
systemZone, that can figure out the right offset based on the date directly.

Also, fix the test to report durations for the calculated timeouts.