Drbd,min-io: Restart DRBD resource if blockSize changed
If existing peers have a given block-size and a new peer with a larger
block-size we must restart all peers so that everyone uses the new larger
block-size.
This can only be done if all DRBD peers are secondary since the application on
top of a primary cannot be updated.