SatelliteMigrations: added first version
This first satellite migration needs to set ZFS properties on ZFS snapshots that LINSTOR
created during "linstor rd clone ... --use-zfs-clone". The old code that would have
cleaned up those ZFS snapshot once the cloned resource is deleted, was removed. The new
method is that if LINSTOR deletes a ZVOL or a ZFS snapshot it checks if it has a parent
(i.e. origin snapshot in case of a ZVOL or a base-ZVOL in case of a ZFS snapshot) that
also has the ZFS property "linstor:marked_for_deletion" set to "true". If so the parent
objects are also deleted by LINSTOR, recursively. Base-ZVOLs are of course only deleted
if they have no more clones left.