StltSnapApiCallHandler: Deleting last snapshot now properly cleans up RD and RG
In a scenario where we have a resource on our local satellite (which also means
that we also have the corresponding RD, RG, VD, V, etc objects) and we create
a snapshot from our resource, delete the resource afterwards (this is allowed
with LVM and since the recent rework also with ZFS) leaving only the snapshot
behind that is still preventing the RD, VD, RG (VG) objects from getting cleaned
up. If that final snapshot is also deleted, only the snapshot-definition was
deleting itself and deregistering itself from the RD. The RD was not checked
if it can be deleted or not. Therefore also the RG was never properly deleted.

Leaving the RD behind if the last snapshot was deleted is an issue, because now
the controller has no more use for the RD and allows the user to delete the RD
entirely from the database. The satellite will not be notified about this, since
the satellite also has no longer any reference to the RD.
Since the RD is deleted, the user is now allowed to create a new RD with the
same name as previously. If a resource from this new RD (with old resource-name)
now gets deployed on our satellite, the satellite will find the old RD, causing
a UUID mismatch.

This patch combines the cleanup code from the StltSnapshotApiCallHandler and the
DeviceHandlerImpl.