AbsStorageProvider: Fix finding snapshot layer data during restore
Let's consider the following scenario:
* Resource "source" has a snapshot "snap"
* Snapshot source/snap should be restored into a new resource "target"

The method getSnapVlmData gets two parameters: target's volume-data and the
snapshot name ("snap" in our example) and is expected to return source's
snapshot-volume-data. From the target volume we can get the name of the
source-resource and the snapshot name.
The method tried to find the source-snapshot-volume-data as follows:
target-volume-data -> target-volume -> target-resource -> target-node ->
source-resource -> source-resource-definition -> source-snapshot ->
source-snapshot-volume -> source-snapshot-volume-data.

Besides being cumbersome, this had the issue that we are allowed to delete a
resource while it still has snapshot(s) (in LVM_THIN this was allowed already
and is additionally allowed in ZFS and ZFS_THIN with recent changes).
If the source-resource is simply missing, the above described chain no longer
works and runs into an exception.

Instead of this chain the AbsStorageProvider now fetches the
source-resource-definition directly from the (satellite-) global
resource-definition-map. The resource might have been deleted, but as long as
at least one snapshot exists, its snapshot-definition prevents the deletion
of the resource-definition.