Crypto: Ignore entries that missing type and/or driver
For some reasons we got NullPointerExceptions when trying to serialize a
the ProcCryptoEntry to protobuf while calling .setDriver. I am not fully
sure why /proc/crypto should give us such entries, but this patch now
ignores the entries (instead of causing the FullSync to fail with a
NPE which leads to the satellite hanging in Connected state and never
going into Online state).

This patch might cause other problems if a satellite could end up with
no valid entries at all. For now, we are also logging warnings when
LINSTOR ignores such an entry so that we might find some clues what
about the root cause of this issue.