Drbd,min-io: Prevent resource creation under special circumstances
If a new DRBD peer is added with a higher min-io and the DRBD resource had
previously, all peers need to be restarted (down + adjust) in order for the new
block-size to take effect. This must not be done during any peer (local or
remote) is primary.
Ideally the creation of a new resource on a node which would force an increase
of block-size should be rejected from the controller already due to the
constrained mentioned above. However, since this scenario inherently a
race-condition (a DRBD peer can become primary right after the controller
decided to allow the creation of the new resource) the satellites also have a
similar check right before the "drbdadm down" command is executed. Since that
satellite check also only decreases the probability of the race-condition to
hit instead of completely eliminating, these checks are only a "best effort"
attempts to prevent this situation.

This patch does not only reject the creation of resources under certain
circumstances but also adjusts the autoplacer to ignore storage pool under the
given circumstances.