build: add prop to set java home for dist scripts
This changes the existing property for setting up the startup scripts by
allowing passing in the full path to the java home directory. We no longer
need to hard code the /usr/<something>/jvm/jre-... directories.

Also update the spec file to use the new properties, preparing for building
with Java 21 for RHEL 10.