Drbd: Move deletion of .res files into applyFullSync
This patch fixes the following race-condition:

Applying a new FullSync has generally 2 effects regarding the DeviceManager:
1) All core objects are deleted and recreated
2) The DeviceManager's fullSyncFlag is set to true.
If the DeviceManager thread was "hanging" in one of the phase* methods (i.e.
waiting for responses from the controller) while a new FullSync was applied,
the next (not the current!) DeviceManager-run triggered the deletion of the .res
files. However, the current DeviceManager-run potentially creates the .res file
and remembers this via the DrbdRscData.resFileExists boolean. That means, after
the current DeviceManager run this resFileExists is true. As mentioned earlier
in the next DeviceManager run all .res files are deleted (due to fullSyncFlag
being set to true from the applied FullSync). If LINSTOR believes the .res file
should exist while it does not causes a NoSuchFileException.

This patch prevents this race-condition by simply moving the deletion of all
.res files into the StltApiCallHandler's applyFullSync method where all write-
locks are taken and the core objects are just being destroyed (and recreated).

Additionally, as a second safe-guard, a new catch-block is introduced for
NoSuchFileException such that only a warning is logged instead of an
ErrorReport.