Db,Props: Fix wrong instanceNames
Commit 7cddef55f80fbfbbf7c7884917d92dae9d983b14 renamed the instanceNames.
The database migration itself regarding resource-groups was correct, but during
loading the data the old instanceName was used.
The same is true for resource-properties within snapshots as well as volume-connections.
The latter however is not really a problem since the user currently has no API to
set or modify volume-connection properties.

This patch fixes the issue and also adds new migrations in case new properties
were added to existing resource groups (newly set properties will override
old values). The case of resource-properties within a snapshot is easier since
those cannot have been modified by the user, so a simple renaming of the
instanceName should be enough.