Backup,ship: Allow fallback to full backup if node unavailable
When issuing a 'backup ship' command, an incremental backup is
the default option. For that to happen, the node that did the
previous shipment needs to be available (otherwise, drbd-metadata
gets scrambled). If that node is offline or otherwise unusable
the shipment process used to simply fail with a completely inaccurate
error message. Now, this case will simply lead to a full backup being
made instead on another available node, and only if there is no
eligible node for the shipment at all will the command fail.