AutoDiskfulTask: Introduce AutoUnplacer
This new component should decide which of the existing resources should be deleted.

When the a primary diskless gets toggle-disk-ed this new placement might violate existing
resource-group settings. Another way to violate the settings is by simple user interaction
i.e. manually placing multiple resource on the same "site" while having
"--replicas-on-different site" configured on the resource-group.

If any component (AutoDiskfulTask, BalanceResourceTask, etc) should delete an existing
resource, the AutoUnplacer now selects the resource with the most violations. In case of a
draw the storage pools are compared, using the same algorithm as the AutoPlacer but instead
of selecting the storage pool with the highest score (i.e. "most free space", "fewest
deployed resource", etc), it selects the storage pool with the lowest score (i.e. "least
free space", "most deployed resources", etc).