DrbdVersion: ensure all access is synchronized
Should there be two reconnect attempts running at the same time, there
might be two calls to "DrbdVersion.checkVersions()" in parallel. While access
to this method is synchronized, later usage of the DrbdVersion object was not.

This meant that the internal call to "DrbdVersion.restoreDefaults()" might
cause the result of another thread to suddenly report unsupported DRBD
versions.

To fix this, ensure all public access goes through the same synchronized
section.