Drbd,min-io: Suppress DRBD events when briefly restarting DRBD resource
We need to suppress the destroy event of the DRBD peer since this event would
close our internal stream of events for this DRBD resource. A closed stream is
processed by the satellite as well as by the controller. However, since we know
that we will very soon start the DRBD resource again, new events will attempted
to be sent to the controller. But if the controller is no longer expecting
events, newly received events will cause issues on the controller.

Re-initializing the stream would be an option, but that might be more
complicated/cumbersome than to simply ignore a single "destroy resource" event.