BackupShipping: start queue immediately if sending failed
Usually for an l2l-shipment to count as completed the src-ctrl needs to
receive the "requestReceiveDone" from the dst-ctrl. Only then does the
src-ctrl try to start further shipments from the queue. However, if the
src-stlt fails to send (for any reason), there is no reason to wait for
the confirmation that the receive also failed - in fact that confirmation
might never arrive because the fail might have been due to the dst-ctrl
going offline. Therefore we try to start queued shipments if the success
boolean is set to false.