BackupShipping: clean up queues on snapshot delete
When a snapshot gets deleted, a check exists to make sure it isn't queued.
However, the snap could still be set as the base snap for an incremental
shipment in the queue - in that case, the queue-items that have this snap
set now get their prevSnap set to null, forcing a full backup.