Rework auto-quorum behaviour
Remove the auto-quorum property and its behaviour of copying the current
property value into the resource-definition.
Now we track if the quorum drbd option was set by the use and if so,
leave the property as it is and only work on resource definitions,
that are not altered by the user.