Reconnect: Fix double reconnecting race condition
When an explicit "linstor node reconnect" is issued, the controller first closes
the current connection and afterwards issues a new connection attempt.

While the connection is already closed but before the controller can start the
new connection attempt, there was a small time window where the PingTask could
have still a reference to the peer (and therefore to the already closed
connection). When the PingTask sees a peer that is not properly connected, it
issues a reconnect.

In the worst case this could end up having 2 connections to the same satellite.
The satellite drops the older connection by closing it and wants to keep the
newer connection. The controller however then sees that one of its connection
was "unexpectedly" closed, so it issues a reconnect resulting again in having
2 connections to the same satellite. This can repeat indefinitely.

This patch fixes the issue by simply marking the peer to not allow a reconnect
before the connection is closed. This patch does _not_ prevent the PingTask to
iterate over a peer with a closed connection, but rather the PingTask rechecks
the peer if a reconnect is allowed or not. If the reconnect is not allowed, the
PingTask silently drops the peer from its ping-list.