DRBD: Fixed orphand .res file
The .res file was only deleted if the DRBD resource was up when LINSTOR tries to delete it.
If the DRBD resource was already down, LINSTOR would result in a no-op. This no-op also
skips the check / attempt to delete the .res file, possibly leaving the .res file on the
system.

This could result in a "duplicated minor number" error from drbdadm when LINSTOR afterwards
reuses the freed up minor number and generates a new .res files for a different resource.
drbdadm currently still scans all existing .res files and finds the minor number in the
orphaned .res file as well as in the just generated .res file.