ProcCryptoTest: Properly close errorReporter after test
The test creates a new StdErrorReporter, which by default connects to the H2
error database. However, before this patch the test did not call shutdown which
lead to a few confusions within the H2 drivers.
The test itself did not notice the side effects, but not closing the H2 database
properly causes the next test to reuse the connection.

This did go unnoticed because of the somewhat random order of the unit tests in
general. If ProcCryptoUtilsTest is run before a test that checks for error
report count, like DBLoggingTest, the latter test will fail due to unexpected
error reports.

This patch had to convert the "... extends TestCase" into the @Test annotation
based since the former approach does not allow for @AfterClass usages.