PingTask: Fixed rare "access to deleted node"
The PingTask did not take any locks while iterating through the peers.
In order to prevent pinging the same node through multiple peers, the PingTask
queries "peer.getNode().getPeer()". If those two peers are not identical, one
is dropped from the PingTask.

The issue now was that ".getPeer()" could cause an "Access to deleted node"
exception when the node was deleted while the PingTask tried to run the ping.

Note: Removing the peer first from the PingTask before actually calling
node.delete() would not have solved the issue, since the PingTask could have
already cached the list of peers.

This patch fixes the issue by performing the identity check of the peers
while holding the readlock.