AppLifeCycleMgr: Fixed deadlock during shutdown
Previously the ApplicationLifeCycleManager took the reconfigurationWriteLock
for not just setting the shutdown flags for every service but also while waiting
for the services to properly shutdown.
If a service woke up and tried to acquire any reconfiguration*Lock before it
initialized it it's shutdown but the reconfigurationWriteLock is already taken
by the ApplicationLifeCycleManager, we had a classical deadlock.

This patch fixes this issue by only taking the reconfigurationWriteLock for
setting / initializing the shutdown of all services. The waiting for the completion
of the shutdown procedure is done without holding the lock.