Prevent VlmDfn UUID mismatch
When deleting a rsc, the stlt assumed that the corresponding vlms
and vlmDfns are deleted automatically. This, however, is not the
case if a snapshot of the rsc exists. Therefore the vlms need to
be deleted separately to prevent the vlmDfn from continuing to
exist on the stlt, which will lead to a UUID mismatch if a vlmDfn
with the same vlmNr as the deleted vlmDfn is created later on.