Ensure zfs-vlm-suffix is set correctly during vlmDfn delete
When deleting a rsc or a vlmDfn of a rscDfn that has snapshots,
the zfs-vlms cannot simply be deleted due to how zfs-snapshots work.
Therefore the zfs-vlms are renamed to prevent linstor from assuming
they belong to existing rscs. The suffix for that renaming is decided
by the ctrl, but was not set during a vlmDfn delete, leading to
wrong zfs-vlm-names. This is now fixed.