LayerIgnoreReason: fix problems with ignore-reason and skip-disk
Removed skipping layers if the ignore-reason didn't change, since
it doesn't really bring much benefit but makes some special cases
(such as re-creating part of the layer-tree during toggle disk)
more complicated.

Set the exists-boolean of the vlmData to false after trying and
failing to remove a vlm because skipDisk is set.