drbdlayer: optimize drbdadm adjust call by asking drbd for changes
In the prepare step, we use the new drbdadm command sh-list-adjustable
to get a list of all resource that need to be adjusted.
Before that we have to generate the .res file so drbd knows about the changes.
If .res file generation isn't possible or has a problem, we will
adjust the resource by default.