Clone,ZFS,copy: Fix race condition during cleanup
Previously the satellite ran as one command something like this:
  zfs send ... | zfs receive ... && zfs destroy ${cloned_rsc}@%;
And addtionally deleted the source-snapshot that was used in the "zfs send" part.
However, both of these "zfs destroy" operations were prone to "device is busy"
or occasionally even to "could not find any snapshots to destroy" errors.

This patch decouples the ".. && zfs destroy ... " part into a dedicated cleanup
step and uses a newly introduced ZfsUtils.deleteIfExists which tries to delete
the given ZFS identifier while ignoring some "not exists" error messages and
in case of other errors retry with a delay the "zfs destroy" command.