DrbdRscData: Init ports into backing set
The previous commit did the following:
In the constructor of DrbdRscData initPorts was called which tried to reuse code so
it called the private version of setPorts. setPorts on the other hand operated on
the TransactionSet rscPorts. Adding to this set should trigger a database update
eventually, but since usually the parent object (in this case DrbdRscData) already
exists during such a set.add(...) the operation is mapped to an update towards
K8s.
The issue is that this update was triggered from the constructor of the DrbdRscData
which itself was not created yet. That means that the K8sCrdTransaction was populated
first with an update and afterwards with a create operation. The create operation
triggered the ImplementationError:
"Tried to create resource " + k8sRsc.getKind() + "/" + key +
" that was already created or updated in the same transaction"

This patch fixes the issue by bypassing the TransactionSet during the initPorts
call and directly populate the backing set, effectively preventing the database
updates during initialization.