Snap,rollback: Introduce property for old rollback behavior
The current (also refered as "new") mechanic of "linstor snapshot rollback" first deletes (or
renames in case of ZFS) the existing LV and then restores the given snapshot into an LV with the
original name of the resource.

This is still necessary for LVM due to an LVM snapshot losing its origin if a different snapshot
with the same origin gets "merged" into its origin.
Although this works on ZFS this new mechanic has 2 downsides:
1) Since the "_deleted_${rsc}_${timestamp}" ZVOL needs to be kept until a snaphot or a ZFS clone
of the snapshot exists (and the result of a "linstor snapshot rollback" is indeed a ZFS clone of
the given snapshot) an increasingly long chain of snapshot + clone + snapshot + clone + ...
needs to be kept until the last dependent snapshot or ZVOL is finally deleted. This can easily be
very space-consuming.
2) The fact that LINSTOR runs a "zfs clone ... " as the second part of the "rollback via restore"
procedure, the resulting ZVOL after a "linstor snapshot rollback" is technically a different ZVOL
than the original in terms of ZFS internal GID. This can easily lead to problems when trying to
ship ZVOLs to some external place (or re-apply the sent ZFS snapshots)

This patch introduces a new property: "StorDriver/Zfs/RollbackStrategy" which currently can have
one of three values:
* "InPlace": This is the old mechanic where LINSTOR will run a "zfs rollback ...". This also re-
introduces the limitations that come from ZFS, for example that only the most recent snapshot can
be used as a rollback target. Intermediate snapshots need to be deleted manually.
* "Restore": This is the new mechanic as described above. In short: Delete/Rename original LV and
use "linstor snapshot resource restore ..."
* "Dynamic": If "InPlace" strategy seems applicable from the controllers point of view "InPlace"
is chosen, "Restore" otherwise.

The default value is "Dynamic" as an attempt to avoid unnecessarily wasting space.