Zfs,delete: Add strategy to switch from delete vs rename
A ZVOL cannot be deleted while it has snapshots. Since v1.31.0 LINSTOR renamed
the ZVOL to "_deleted_{orig_zvol}_{date_time_of_deletion}" to free up the name
so a new resource (ZVOL) can be created instead of the just deleted.

Snapshots are also renamed accordingly on ZFS level.

However, this could lead to unexpected behavior. This patch re-introduces the old
behavior that the deletion of a LINSTOR resource is rejected by the controller if
the controller knows about snapshots of that resource.
This behavior can be selected via "StorDriver/Zfs/DeleteStrategy". The options are
"Delete" (fail if snapshots exist for ZFS), "Rename" and "Dynamic". "Dynamic" runs
"Delete" if no snapshot exist and "Rename" otherwise.
The default is "Dynamic".