Backup,Ship: Set SnapVlmDfn size
Without setting the size the SnapVlmDfn is created based on the current VlmDfn's
size. This could mean that the current VD has for example 10G but we are about
to receive 20G of backup. The receiving works just fine, but if the SVD stores
10G as size a rollback or restore will set those 10G as the new VD size. This
will cause an issue when a new peer joins that creates a new resource with the
(too small) VD's size and DRBD complains about "peer disk is too small".