Drbd: Send existing connections after stlt reconnect
If the satellite gets a new connection from the controller, the cached DRBD
connections were not sent to the controller.
Regardless if the satellite was restarted or not, the events from
"drbdsetup events2" stream are processed before LINSTORs FullSync. That
means that we have no chance to send the information to the controller
while processing the events.

Therefore this patch re-sends the connection events now to the controller.