S3: Reduce multipart max part count from 10k to 1k
Some S3 implementations do not support 10k parts. Although Amazon itself does support 10.000 parts
every part can only be up to 5GiB and the sum only 5TiB. So even with only 1.000 parts we are still
not limiting the size of an uploaded object to AWS.

A future commit should make the part-count, part-max-size and part-min-size configurable.