Drbd: Adjust optimizations fixes
The long term "FIXME" (that we always set the requiresAdjust to true since we could not figure
out when the SP's PrefNic was changed) was removed since the new method using
"drbdadm sh-list-adjustable" already requires us to regenerate all .res files. If the PrefNic
was changed it would result in a different .res file which triggers drbdadm to show the resource
in "sh-list-adjustable".

Also fixed that if neither RESIZE / DRBD_RESIZE flags are set we not necessarily need an adjust
(whereas if both are set we definitely want to adjust)