LayerMerger: Replace root layer object during toggle-disk
The LayerMerger already threw away sub-trees of the layer data if they changed
and were newly created (which only happened on the satellite, never on the
controller). However that was only true for the sub-trees but not the root of
the layer tree. That means that even if the controller threw away the entire
layer tree during a toggle disk the satellite still kept the old root node.
That means for example DRBD (ID 1) with a child Storage (ID 0, with LvmData) was
replaced with DRBD (3) -> Storage (2, with DisklessData) was "merged" on the
satellite as DRBD (1) -> Storage (2, with DisklessData).
Essentially the structure was correct, which is the reason this mistake was
never detected and also never caused any harm.

Now using the "MismatchIdMergeStrategy" implemented recently this mistake
gets detected and causes issues.

Since the satellite always uses FORCE_MERGE strategy, the satellite will always
just replace its local data with what the controller is sending if the IDs
mismatch. The controller currently should not use FORCE_MERGE since that would
mean that the satellite is allowed to change the layer tree on the controller
(usually this is the other way around). Future implementations / features /
layers might change this. For now the controller should only be asked for a
MismatchIdMergeStrategy if the IDs differ which should only be possible during
a toggle-disk event when the satellite still send outdated data. This case
should however be handled by registering the old -> new ID during toggle disk.

If the controller does not detect such a case, an exception is thrown (just as
the "Received unknown Storage data from satellite" exception from before this
patch-series)