BackupInfo: Introduce waitForShipDoneFlux
This new method returns a flux that will be completed or flux-errored when the
corresponding shipment is done. If the shipment was successful the new flux
will be completed, otherwise a new ShipmentFailedException is added to the flux.

The intention of this new flux is that components can now register with
post-ship-hooks on the controller to run other fluxes. It is highly recommended
to put this new flux into a background runner, otherwise the client will also
be delayed until the shipment is finished (not just started).