CopySnapHelper: Add methods to migrate snapshot
The intention of these new methods is to ship one existing snapshot to a new
node and once the shipment is done to properly delete the source snapshot.
This will be necessary for an upcoming "node ecavuate" improvement.

This new method starts a background-flux based on the recently introduced
waitForShipDoneFlux.