CopySnapHelper: Search for local remotes
Instead of always creating the "local-remote-generated-by-linstor", check if
an existing LinstorRemote's cluster-id is the same as our local cluster. If so
take that LinstorRemote.

There are two reasons for this:
1) prevent unnecessarily creating a LinstorRemote
2) (more important!) a LinstorRemote that was created by the user has a higher
chance to also contain a passphrase. The generated LinstorRemote cannot know the
local controller passphrase since that passphrase is never stored (only the
encrypted masterKey, but l2l shipment requires the passphrase not the masterKey)