Backup,create: Fixed bugs when shipping existing snapshots
This patch contains several bug fixes:

* Methods like "postCreateSnapshot" must not be called if the snapshot to be
shipped already existed.

* Properties fixed:
** The key containing the source-node's name must not contain the target
resource name. The reason is simply because the source-satellite does not have
the information and therefor cannot properly find the property which states if
the local satellite is the source-satellite or not.
** If a snapshot is part of a chain of incremental snapshots (i.e. in case of
S3) the snapshot had as a source property (to the same remote we are now
restoring to) the KEY_BACKUP_LAST_SNAPSHOT property set so that the satellite
that did the sending previously knew on what previous snapshot the incremental
shipment should be based on.
If during a restore such an old source property was set to the same remote, we
need to copy that property into the target backup namespace so that the
receiving satellite knows that it needs to create a new snapshot based on a
specific previous snapshot in order to restore the incremental chain.
** When setting the shipping status to "Shipping" right before a restore, we
also need to set the KEY_BACKUP_DST_NODE property so that the satellite can find
itself in the value so that the local satellite knows that it is the one that
should do the receiving, not an other satellite that also received this update.
** Source properties must not be deleted when restoring a target since we
(currently) do not know if those properties are from an active shipment or were
downloaded from S3.