Backup: Introduce Snapshot.shipBackup + new property
Re-using Snapshot.takeSnapshot for starting shipment could cause issues with the
new "ship existing snapshot" feature. One scenario is a node evacuation where
in the same cluster, in the same SnapshotDefinition the evacuation target node
needs to prepare / create a snapshot (hence the actual .takeSnapshot usage)
while a different node is waiting for the very same .takeSnapshot for
starting the shipment as shipping source. Since we usually set the .takeSnapshot
on all nodes of a given SnapshotDefinition, this might cause a premature start
of the shipment.

A very similar issue occurs if the timing is a bit different such that the
evacuation target finishes creating the snapshot earlier than the shipping
source processed the .takeSnapshot. That could cause that the "I am done
creating the snapshot and am ready to receive" deactivates the .takeSnapshot
that the shipping-source needs to start the shipment. This scenario leads to
the issue that the shipment is simply never started.

The newly introduced .shipBackup boolean not just decouples the two
functionalities from each other but is also only set on single snapshots instead
of all snapshots of a given SnapshotDefinition.

--------

Additional to the above described Snapshot.shipBackup this commit also
introduces a new target property KEY_BACKUP_DST_NODE so that a satellite can
check if the local satellite is the one that should start the receiving.