Shipping: Fixed race condition
The method startDaemon is updating some maps of the ShippingInfo. The other
users of those maps synchronize on "snap" object. So far this was not an issue
since a snapshot could only be used once as shipment which ensured that
startDaemon is run before any other users even started. However since we now
can ship existing snapshots as well we can no longer rely on this assumption.