StorPool: Also send remote storPools to Stlts
For some reason a satellite could until now only request its own storage pool
from the controller.
This was a problem if a property like PrefNic is set on a remote storage pool
but our local satellite needs to configure its local DRBD resources to use the
remote storage pool's configured PrefNic.

This was only partially an issue until now because of the verbose FullSync
packet. All volumes contained information about the storage pools they were
located in. This still did not allow a satellite to request remote storage pools
but if the satellite is restarted or reconnected, it received the otherwise
missing information about remote storage pools.

This patch fixes the issue so that every satellite can now request local and
remote storage pools. The controller is also updating now all effected
satellites about changes in a storage pool, including local as well as remote
storage pools.