physical-storage: pass arguments to pv/vg/lv/zpool create
A user might want to set specific options when creating a storage pool from
plain disks, such as:

* metadatasize for pvcreate
* physicalextentsize for vgcreate
* stripes and stripe size for lvcreate
* failmode for zpool create

These options can generally only be set on creation, so it does not make sense
to somehow store them on the Storage Pools.

For validation, we assume the user provides valid arguments, as we cannot
easily verify that all arguments are valid in the required command lines.