Backups: prevent race when auto-deleting snapshot after shipping
The copy-snapshot feature that is used during an evacuate wants
to delete the copied snapshots once they have been shipped. If
the delete triggers too soon, it could happen before the receive
was completely done, leading to an error for the delete that the
src-snap can't be deleted during shipping. Now the delete is only
triggered once the receive is actually done, preventing this race.