StltStorPool: Fix hanging Stlt
The recently added patch 62765556fc96f08a9b1915f0602479ca9051919b skips applying the storPoolPojo
if on the local satellite the storPoolPojo's node does not exist. This is fine since it might be
a race-condition where the controller already sent us a "ChangedStorPool" of a remote storPool while
our local devMgr was still working on a resource. If that (remote) resource was deleted and it
was the last resource of the given remote node, the devMgr also deletes the remote node, including
the remote storage pool. If the devMgr finishes with that cycle, it will find the "ChangedStorPool"
message and starts to process it. Hence the "ignore unkown remote storage pool".

However, the patch made the mistake to pass an empty storPoolSet (i.e. a set of storPools that had
been applied) to the devMgr, so the devMgr kept waiting until all updates were applied before starting
the actual devMgrRun. This led to a hanging satellite, and since the satellite did not continue, also
the controller's flux got stuck, which can cause client-commands to get stuck.