DmSetupUtils: Make more modular
The purpose of this restructuring of the DmSetupUtils class is to make it easier
to reuse parts of the methods. An upcoming patch will need to not just use
"dmsetup create ... --table $table" but will also need to run
"dmsetup reload ... --table $table" with the same content for $table but also
with partially different content.