Msg: Increase size 16M -> 64M
Some users are already running into this issue especially during FullSync.
If the FullSync gets larger than 16M (yes, the controller can create
packets larget than the limit) the sateliltes simply do not properly
parse the FullSync packet. This leads to the status Offline (FULL SYNC FAILED)
issue.

This patch can (should?) be reverted once we rework the FullSync protocol
or the FullSync message. That packet needs to get smaller on the one hand
but the controller (and satellite of course) need to be able to process
multiple FullSync (-parts) so that we can always stay under the current
limit of the message size.
Once that is done, this patch gets obsolete.