Drbd,min-io: Add database migrations
This migration deactivates this new auto-block-size feature for all existing resource-
definitions. Resources might need to be restarted if the block-size changes, and since
we cannot reliably calculate what the block-sizes on devices were during this migration
it would be too dangerous to calculate a wrong value here.

That means that this min-io feature will only be active for newly created resource-
definitions or if the user explicitly enables it for existing RDs.