Drbd,min-io: Add auto-/ freeze-block-size
When a DRBD resource becomes primary LINSTOR will automatically set the freeze-block-size
to True on the volume-definition. That property can be unset by the user again or will
be unset automatically when the last resource of the resource-definition is deleted.

If the freeze-block-size is True, the block-size must not be changed regardless if there
is a primary peer or not.
The reason for this that some tools (like XFS) do not handle very well if the blck-size
changes (i.e. from 512 -> 4096) from one usage (mount) to the next