Clone,Autoplacer: Fix wrongly calculated min-io size
When a base resource is cloned the controller already created the target
RD. With this empty RD (i.e. no resources) the autoplacer was executed
to select storage pools for this new RD.

The min-io calculations wrongfully claimed to want to have 512 min-io
even if the source RD as well as all stroage pools had 4096 min-io size.

This patch fixes the issue in two ways:
1) The ResourceDefinition.getFloorVolumesMinIoSize now returns null if no
resource exists
2) CtrlRscStateHandler.canChangeMinIoSize now also returns true if the
given RD has no resources.

Regarding 1) the autoplacer was also adjusted to only consider the min-io
if it is non-null. Both ways should fix the issue.